Found out that long int is actually 64 bit instead of 32 (I thought that was what long long was for), so replaced everything with uint32_t; added def(), set(), if() and while() functions to rvd, as well as a couple more operators, to make it a proper (if simple) functional language; cleaned up map file syntax a little and fixed docs here and there.

+++ debug-interface/src/rvd/definitions.c
@@ -64,7 +64,7 @@
  * function.
  */
 static int getBin(const unsigned char *str) {
-  unsigned long hash = 0;
+  uint32_t hash = 0;
   int c;
   
   while (c = *str++) {
